fix hardlink behavior when extracting layers
extracting in reverse order is more optimal, but it makes supporting
hardlinks very difficult. the 'concourse/git-resource' image didn't work
because a hardlinked file was removed in a later layer, making it
impossible to link to in the earlier layer.

it may be worth investigating another solution; it's really nice to not
have to write files that are later removed.