out: fix uploading img save'd tarballs
for some reason this was failing with 'blob sha256:... not found', where
the digest corresponded to the config blob. i noticed containerregistry
handles this properly for uncompressed images, but not compressed
images, which img saves as.

this may be something to actually fix and PR to go-containerregistry in
the future.