fix tag defaulting
previously the default tag would only be applied if the tag was
explicitly provided as an empty string, because the default was applied
during parsing. however if the field was not specified at all, nothing
would be parsed.

now the default is applied when the tag is accessed on the object, and
works even if the field was completely omitted.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>