don't use retryable transport
injecting our own retryable transport seems to result in the following
errors:

* DIGEST_INVALID: this is surprising, but is the most common failure;
  something must be partially reading or otherwise butchering the payload
* UNAUTHORIZED: perhaps tokens expire in the middle of the transport
  retry loop? the official retry logic seems to handle this fine.
* BLOB_UPLOAD_UNKNOWN: perhaps from retrying when the upload actually
  succeeded server-side?

using the existing go-containerregistry retry logic seems to be more
durable.

we'll still need some way to retry 429 errors, but that doesn't need to
be all the way down at the transport layer, since that's an error from
the entire service, not an individual request. for this we can just have
an outer retry loop.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>