check: respect 'from' version
note: we are following the future-looking resource check semantics here.
the initial check will return all versions. later checks will treat the
tag and digest as a cursor: as long as the digest is the same, only the
versions after the tag will be returned. if the digests differ, all
versions will be returned instead, "resetting" the history.

Signed-off-by: Alex Suraci <asuraci@vmware.com>
Co-authored-by: Bishoy Youssef <byoussef@vmware.com>