check: re-use transport to prevent repeated auth
this will perform a ping + auth flow and then pass the transport along
in options for all subsequent requests.

the net effect is that rather than pinging + acquiring a token on every
request, we will only ping, since the token is already acquired. i'm not
sure if either of these endpoints are rate limited in the first place,
so this might not matter, but it at least saves time.

Signed-off-by: Alex Suraci <asuraci@vmware.com>