Call out that aws_region is required for ECR
It wasn't clear to me that you had to include an `aws_region` if you wanted this resource to know to talk to an AWS ECR repository. Maybe this was dumb user error on my part but without an example of how this would work with ECR I figure this might help others too.

Signed-off-by: Matt Dodge <mattedgod@gmail.com>