feat(credentials): Use default key chain for auth
This should allow simple specialisation of the image to allow for specific container registry variants.

For example, to use this with GCR the existing registry-image image can be used as a base layer and then we can copy in a ~/.docker/config.json and optionally a credential manager to authenticate.