properly handle whiteout opaque dirs when unpacking
concourse/registry-image-resource#267

If a file in the layer tarball is labeled foo/.wh..wh..opq, that indicates that
the existing contents of the directory foo/ should be emptied. We were
previously treating this the same as normal whiteout files (e.g. foo/.wh.myfile
means delete foo/myfile), which doesn't work properly.

With this commit, we're now able to delete a directory and recreate it in the
same layer, and have it be unpacked properly.

Signed-off-by: Esteban Foronda <eforonda@vmware.com>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>