fix: support OCI Image Layout Specification
This adds "oci-layout" as a format during "get" which saves out the
contents per the OCI Image Layout Specification.

It also fixes the corresponding "put" to upload all images (including
multi-architecture images) that are written according to this
specification.

This handles both modern image indexes, as well as legacy image
references.

Amongst other things, this makes this resource type more suitable for
mirroring images from one repo to another, as this now will (a) preserve
the digest and (b) mirror all images for each architecture in the source
image.

Signed-off-by: Adam Eijdenberg <adam.eijdenberg@defence.gov.au>