update to aws-sdk-go-v2
The v1 SDK is reaching EOL this year.

The big thing that got removed here as part of the update is the wrapper
we had around GetECRAuthorizationToken(), which added RegistryIds to the
scope of the token to enable folks to do cross-account ECR requests.

As per the comment in the aws Go SDK, this is no longer required. They
state that the token you get has access to all ECR's that the underlying
IAM principal has access to.

Signed-off-by: Taylor Silva <dev@taydev.net>