unwrap nested Image Indexes
I found that when pushing a multi-arch image to Docker Hub, I wouldn't
see which platforms the image supported like I would when pushing a
multi-arch image with the Docker CLI.

After digging into the OCI-layout I found out that buildkit, when
exporting the image to disk in oci-layout format, wraps the image index
in another image index. I'm not sure why it does that; haven't bothered
to dig into it in the buildkit code. It seems buildkit only does this
when writing the image to disk. The ImageIndex wrapped looks like this:

```json
{
  "schemaVersion": 2,
  "mediaType": "application/vnd.oci.image.index.v1+json",
  "manifests": [
    {
      "mediaType": "application/vnd.oci.image.index.v1+json",
      "digest": "sha256:55bf56746c776b4cc84498657dbaeeb33b247956dc08c6763b6ce2a7b53d2da2",
      "size": 647,
      "annotations": {
        "org.opencontainers.image.created": "2025-04-22T21:01:52Z"
      }
    }
  ]
}
```

The problem is that when you push this wrapped ImageIndex, image
registries like Docker Hub don't know or display the platforms the image
supports.

When pushing the image using `docker push` or using the oras CLI to
push the local oci-layout image I made from buildkit, I found they both
pushed the nested image index instead of the parent. The nested image
index looks like this btw:

```json
{
  "schemaVersion": 2,
  "mediaType": "application/vnd.oci.image.index.v1+json",
  "manifests": [
    {
      "mediaType": "application/vnd.oci.image.manifest.v1+json",
      "digest": "sha256:4fce054b00b7a0363412c153bfec7e2093fd5dbd3a64256f21c1843bbc35f98e",
      "size": 669,
      "platform": {
        "architecture": "amd64",
        "os": "linux"
      }
    },
    {
      "mediaType": "application/vnd.oci.image.manifest.v1+json",
      "digest": "sha256:74b79ed32d2ba388471560515c3be5980c89d5143ae760a7b9a005e2b925907c",
      "size": 668,
      "platform": {
        "architecture": "arm64",
        "os": "linux"
      }
    }
  ]
}
```

This commit has the registry image unwrap all nested image indexes until
it gets an ImageIndex that contains more than one manifest that is not
another ImageIndex.

I've tested this and pushing multi-arch images with the registry-image
now correctly displays the platforms that were pushed. When mirroring
images, the original digests are still preserved as well.

Signed-off-by: Taylor Silva <dev@taydev.net>